address feedback pt 1
