+++ packages/manager/src/features/PlacementGroups/PlacementGroupsDeleteModal.tsx
@@ -116,7 +116,7 @@ export const PlacementGroupsDeleteModal = (props: Props) => {
       onClick={onDelete}
       onClose={handleClose}
       open={open}
-      title={`Delete Placement Group${selectedPlacementGroup ? ` ${selectedPlacementGroup.label}` : ''}`}
+      title={`Delete Placement Group${selectedPlacementGroup ? ` ${selectedPlacementGroup.label}` : ''}?`}
     >
       {error && (
         <Notice
